feat: implement dead zone for controller
This change adds a dead zone on either side of neutral for each analog
stick. In testing, we found that the default neutral position for the
left analog stick was not truly at 128, and not having a dead zone did
not provide enough tolerance to let turning in place work normally.

Did the same thing for left and right, because a dead zone is just a
helpful practice in general.